before
[denis@f22 helix]$ ./order_book_perf_test 
order_book::add()    61 ns/op
order_book::cancel() 10 ns/op
order_book::remove() 32 ns/op

after
[denis@f22 helix]$ ./order_book_perf_test 
order_book::add()    48 ns/op
order_book::cancel() 10 ns/op
order_book::remove() 19 ns/op