@penberg 2015-08-23
@divaykin Can you please add

```
sudo apt-get install libboost-all-dev
```

in the ``before_install`` section of ``.travis.yml`` to fix Travis CI build?