@divaykin 2015-08-23
seems like boost version < 1.55 installed. i.e. no emplace method. what's the best way to get a new version? should we get it from source?