@penberg 2015-08-23
@divaykin Yeah, it's probably easiest to try to install from sources like we do with ``libuv``.