@divaykin 2015-08-23
@penberg does it worth the hassle? do you see the same improvement i see / like the code?