@penberg 2015-08-23
What CPU do you have? Which GCC version and Boost version are you testing with? I'm using a Haswell and running Fedora 21 which has:

* gcc version 4.9.2 20150212 (Red Hat 4.9.2-6) (GCC) 
* boost-devel 1.55.0

Order book performance is *worse* with your changes:

```
Before:

[penberg@nero helix]$ taskset -c 1 ./tests/order_book_perf_test 
order_book::add()    54 ns/op
order_book::cancel() 9 ns/op
order_book::remove() 28 ns/op

After:

[penberg@nero helix]$ taskset -c 1 ./tests/order_book_perf_test 
order_book::add()    68 ns/op
order_book::cancel() 18 ns/op
order_book::remove() 22 ns/op
```