@divaykin 2015-08-23
gcc version 5.1.1 20150618 (Red Hat 5.1.1-4) (GCC)
boost 1.57.0

fc22. CPU is an old one, tomorrow will try on Broadwell also.