@divaykin 2015-08-23
could you please try the newer software if possible? thx @penberg