@penberg 2015-08-23
@divaykin I can try newer versions but if performance is better on Broadwell I think we can just merge this.