@divaykin 2015-08-23
i'd very much appreciate independant benchmark, maybe there is something wrong i'm doing