@divaykin 2015-08-23
i think boost 1.56 had really a lot of changes to Multi Index see http://www.boost.org/doc/libs/1_59_0/libs/multi_index/doc/release_notes.html probably just getting newer boost should be sufficient