@divaykin 2015-08-24
```
Intel(R) Core(TM) i7-5600U CPU @ 2.60GHz
before:
[denis@x1 build]$ ./tests/order_book_perf_test 
order_book::add()    64 ns/op
order_book::cancel() 12 ns/op
order_book::remove() 34 ns/op

after:
[denis@x1 build]$ taskset -c 2 ./tests/order_book_perf_test 
order_book::add()    49 ns/op
order_book::cancel() 11 ns/op
order_book::remove() 20 ns/op
```

would be great if you could test with latest boost @penberg  thx!