@penberg 2015-08-25
Yes, it indeed is a Boost and GCC issue. I ran the test in a Docker container on my Haswell machine:

```
[root@af71f34107bf helix]# cat /etc/fedora-release
Fedora release 22 (Twenty Two)
[root@af71f34107bf helix]# gcc -v
gcc version 5.1.1 20150618 (Red Hat 5.1.1-4) (GCC) 
```

The performance improvement looks similar to yours:

```
Before:

[root@af71f34107bf helix]# tests/order_book_perf_test
order_book::add()    53 ns/op
order_book::cancel() 10 ns/op
order_book::remove() 28 ns/op

After:

[root@af71f34107bf helix]# tests/order_book_perf_test
order_book::add()    37 ns/op
order_book::cancel() 9 ns/op
order_book::remove() 15 ns/op
```

I wonder what to do with Travis CI, though. Building Boost from sources was pretty painful when I tried to do it.