@penberg 2015-08-25
I found these instructions:

https://coderwall.com/p/0atfug/installing-boost-1-55-from-source-on-ubuntu-12-04

Perhaps it's not too painful after all...