@divaykin 2015-08-25
cool thing this multi index eh? ) sure, i'll try this out.

i was also looking for a fast `map` replacement but has not found something decent yet.