@divaykin 2015-08-26
the build is not taking long but should work